requirements.txt
Python-binance
TA-Lib
jumpy
websocket_client
Gspread
GSpread formatting 
